Sketch_to_Reality/requirements.txt
flask
flask-cors
diffusers
transformers
accelerate
Pillow
opencv-python-headless
